Feature/more examples 11 (#3674)
* More SDK testing

* Cleaning

* Captures base.Timestamp for Json parsing

* Cleans fuzzer. Fixes a few json parsing errors. Starts to clean up compressedTx, etc.

* Removed CompressedTx from Slurp type cache - wasn't cached anyway - no upgrade

* Removed CompressedLog and CompressedTrace from caches. Not yet auto-generated

* Auto generates updated field that have been removed

* Finishes fuzzing for chifra slurp

+++ src/apps/chifra/internal/when/handle_ts_show.go
@@ -40,7 +40,7 @@ func (opts *WhenOptions) HandleTimestampsShow() error {
 				s := types.Timestamp{
 					BlockNumber: base.Blknum(ts.Bn),
 					Timestamp:   base.Timestamp(ts.Ts),
-					Diff:        base.Timestamp(ts.Ts) - prev,
+					Diff:        int64(base.Timestamp(ts.Ts) - prev),
 				}
 				if bn == 0 {
 					s.Diff = 0
